core/pyproject.toml
[project]
name = "devteam-core"
version = "0.1.0"
description = "Core services for DevTeam"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "anthropic>=0.75.0",
    "google-genai>=1.53.0",
    "openai>=2.8.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.7",
]
